requirements/base.txt
mysql-connector-python>=8.0.0
PyYAML>=6.0.0
sqlparse>=0.4.4 